refactor: rename some packages and functions
