+++ pkg/cfg.go
@@ -1,7 +1,13 @@
-package internal
+package pkg
+
+import "os"
 
 type Config struct {
 	DatabaseURL string
+}
+
+var MainConfig = Config{
+	DatabaseURL: os.Getenv("DATABASE_URL"),
 }
 
 var TestConfig = Config{
